[PAGOPA-818] kibana integration (#77)
* 🐳 chore(Dockerfile): add Elastic APM agent to Dockerfile
🔧 chore(helm/values-dev.yaml): add Elastic APM configuration to environment variables
The Dockerfile has been updated to add the Elastic APM agent to the container. The agent is downloaded from Maven Central and added to the container as apm-elk-agent.jar. The helm/values-dev.yaml file has been updated to include the Elastic APM configuration in the environment variables. The following variables have been added: ELASTIC_APM_SERVICE_NAME, ELASTIC_APM_APPLICATION_PACKAGES, ELASTIC_APM_SERVER_URL, and ELASTIC_APM_SECRET_TOKEN. These variables are used to configure the Elastic APM agent in the container.

* new version

* fix

* Bump to version 2.5.4-1-kibana [skip ci]

* Bump version [skip ci]

* properties

* Bump to version 2.5.4-2-kibana [skip ci]

* Bump version [skip ci]

* logs

* Bump version [skip ci]

* Bump to version 2.5.4-5-kibana [skip ci]

* Bump version [skip ci]

* 🔍 chore(.gitignore): add helm/charts/ to .gitignore
🔧 chore(values-dev.yaml): change ELASTIC_APM_LOG_SENDING value to string
The `helm/charts/` directory is now added to the `.gitignore` file. This directory is not needed in the repository as it contains generated files. The `ELASTIC_APM_LOG_SENDING` value is now a string to ensure consistency with the other values in the file.

* [PAGOPA-818] 🚀 feat(helm): add Elastic APM configuration to values files Added Elastic APM configuration to both production and UAT values files. The configuration includes the service name, application packages, server URL, log sending, and secret token. This will allow for better monitoring and tracing of the application. 🔥 refactor(application.properties): remove console and file logging patterns The console and file logging patterns have been removed from the application.properties file. This is because the logging configuration has been moved to a separate file for better organization and maintainability. The logging level for the org.springframework package has also been removed as it is not necessary.

* [PAGOPA-818] version

* [PAGOPA-818] updated sonar scanner

---------

Co-authored-by: pagopa-github-bot <github-bot@pagopa.it>
Co-authored-by: pagopa-github-bot <>